"Hide log subscribers & add thirdweb, ethers dependencies"
This commit comments out the lines of code that subscribe to the curator emitter object and log its data for three different topics. The change
is to clean up the codebase as the logging is no longer necessary. Also, three new dependencies have been added to the package.json file: '@thirdweb-dev/chains', '@thirdweb-dev/sdk' and 'ethers'. These added dependencies will allow the application to interact with blockchain networks (namely Ethereum), which is necessary for future updates.