Change fetch method and timestamp handling for origami certification
Modified `fetchLastCertifiedOrigami` to directly call 'latestRoundData' method from the contract, removing unnecessary enforcement and ToDo comment. Also, the timestamp handling in the certification logic has been updated. Now, we're not just checking if there is a previous origami but also checking the timestamp of it. The certification validation has been changed to certify if the last origami is more than 1 hour old instead of the next origami being 1 hour newer. Logging information added to show the remaining hours for origami to be qualified for certification.